requirements.txt
pyyaml>=6.0
pytest>=7.0
pytest-cov>=3.0
pydantic>=1.10.0